pyproject.toml
[tool.poetry]
name = "clarin_centre_registry"
description = "CLARIN Centre Registry Django project configuration"
version = "0.0.0dev"
license = "../LICENSE.txt"
authors = [
    "Beatriz Sanchez Bribian",
    "Michał Gawor <michal@clarin.eu>",
    "Sander Maijers",
    "André Moreira<andre@clarin.eu>",
]
maintainers = [
    "Michał Gawor <michal@clarin.eu>",
    "Andre Moreira <andre@clarin.eu>",
    "<centre-registry@clarin.eu>"
]
packages = [
  { include = "centre_registry_project", from = "Centre-Registry-config" },
  { include = "centre_registry", from = "centre-registry-app" }, 
]

[tool.poetry.urls]
changelog = "https://github.com/clarin-eric/Centre-Registry/blob/master/LICENSE.txt"
documentation = "https://github.com/clarin-eric/Centre-Registry/"
homepage = "https://centres.clarin.eu/"
repository = "https://github.com/clarin-eric/Centre-Registry/"

[tool.poetry.dependencies]
python = '>=3.12.*,<4.0'
Django = '4.2.27'
django-countries = '8.2.0'
django-debug-toolbar = '4.4.6'
djangorestframework = '3.16.1'
django-simple-history = '3.7.0'
django-test-migrations = '1.5.0'
drf_spectacular = '0.29.0'
lxml = '6.0.2'
selenium = '4.40.0'
tzdata = '2025.3'
xmlschema = '4.3.0'

[tool.poetry.group.test.dependencies]
lxml = '6.0.2'
selenium = '4.40.0'
jsonschema = '4.25.1'
django-test-migrations = '1.5.0'


[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
pattern  = "^(?P<base>\\d+\\.\\d+\\.\\d+)(-?((?P<stage>[a-zA-Z]+)\\.?(?P<revision>\\d+)?))?"
format-jinja = """
    {%- if stage == "post" -%}
    	{{ serialize_pep440(base, post=revision) }}
    {%- elif distance == 0 -%}
        {{ serialize_pep440(base, stage, revision) }}	
    {%- else -%}
        {{ serialize_pep440(base, stage, revision, dev=distance, metadata=[commit]) }}
    {%- endif -%}
"""

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
